[RAPTOR-5928] Transform functional test in DRUM
