+++ jenkins/test3_training_model_templates.sh
@@ -50,6 +50,6 @@ pip install -U $DRUM_WHEEL_REAL_PATH
 pip install -r requirements_test.txt
 
 # put tests in this exact order as they build images and as a result jenkins instance may run out of space
-py.test tests/functional/test_training_model_templates.py \
+py.test tests/functional/test_custom_task_templates.py \
         tests/functional/test_drum_push.py \
         --junit-xml="$CDIR/results_drop_in.xml"
