+++ requirements_test.txt
@@ -4,7 +4,8 @@ pytest-runner
 pytest-xdist
 responses
 urllib3 >= 1.25.0
-datarobot==2.24.0
+datarobot==2.26.0b0
+datarobot-bp-workshop==0.1.9.post1
 requests >= 2.24.0
 scikit-learn==0.23.1
 sagemaker-scikit-learn-extension==1.1.0
