+++ tests/fixtures/user_permissions.json
@@ -6,5 +6,7 @@
   "ENABLE_CUSTOM_INFERENCE_MULTICLASS": true,
   "ENABLE_CUSTOM_TRAINING_MULTICLASS": true,
   "ENABLE_CUSTOM_INFERENCE_UNSTRUCTURED_MODEL": true,
-  "EXPERIMENTAL_API_ACCESS": true
+  "EXPERIMENTAL_API_ACCESS": true,
+  "ENABLE_CUSTOMIZABLE_BLUEPRINTS": true,
+  "GRAYBOX_DISABLED": true
 }
